requirements.txt
google-generativeai>=0.3.0
google-api-python-client>=2.150.0
youtube-transcript-api>=0.6.2
python-dotenv>=1.0.0
yt-dlp>=2024.0.0
